컴포넌트화
