+++ 1주차/src/pages/PostList.tsx
@@ -1,9 +1,10 @@
 import styled from "styled-components";
-import BarImg from "../images/Frame.png";
 import "../styles/Text";
 import { Body1, Body2, Title3 } from "../styles/Text";
 import "../font/fonts.css";
 import Post from "../components/Post";
+import Setting from "../components/Setting";
+import NewPost from "../components/NewPost";
 
 const PostList = () => {
   return (
@@ -12,13 +13,8 @@ const PostList = () => {
         <Header>
           <Title>게시물 목록</Title>
           <ButtonDiv>
-            <Setting>
-              <img src={BarImg} />
-              <StyledP>최신순</StyledP>
-            </Setting>
-            <Write>
-              <StyledP2>게시글 작성하기</StyledP2>
-            </Write>
+            <Setting />
+            <NewPost />
           </ButtonDiv>
         </Header>
         <PostDiv>
@@ -38,26 +34,12 @@ const BigDiv = styled.div`
   height: 100%;
   display: flex;
   justify-content: center;
-`
+`;
 const PostDiv = styled.div`
   width: 1200px;
   display: flex;
   flex-direction: column;
   margin-top: 40px;
-`;
-const StyledP2 = styled(Body2)`
-  font-size: 15px;
-  font-weight: 500;
-  color: white;
-`;
-const Write = styled.div`
-  width: 137px;
-  height: 48px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  background-color: #3d8aff;
-  border-radius: 5px;
 `;
 const ButtonDiv = styled.div`
   width: 1200px;
@@ -66,20 +48,6 @@ const ButtonDiv = styled.div`
   justify-content: space-between;
   align-items: center;
   flex-wrap: nowrap;
-`;
-const StyledP = styled(Body1)`
-  color: black;
-`;
-const Setting = styled.button`
-  width: 140px;
-  height: 54px;
-  border: 1px solid #eeeeee;
-  border-radius: 4px;
-  display: flex;
-  align-items: center;
-  padding: 0 25px 0 20px;
-  justify-content: space-between;
-  background-color: white;
 `;
 const Container = styled.div`
   width: 1200px;
